newProject/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
diagrams==0.22.0
Django==4.1.2
emoji==2.1.0
graphviz==0.19.2
idna==3.4
isOdd==0.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
pygame==2.1.2
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tornado==6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
